Added endpoints for all components and made Heroku Ready
To push to heroku:
1. copy build folder from client into public
2. Uncomment final line in gitignore to push up the required keys
