+++ .gitignore
@@ -99,5 +99,5 @@ dist
 # TernJS port file
 .tern-port
 
-# Project Specific
+# # Project Specific
 **/*adminsdk-key.json
