+++ package.json
@@ -11,7 +11,6 @@
     "start": "npx nodemon index"
   },
   "dependencies": {
-    "cors": "^2.8.5",
     "dotenv": "^8.2.0",
     "express": "^4.17.1",
     "firebase": "^8.2.9",
