+++ src/firebaseSetup/firebaseAuth.js
@@ -1,14 +1,8 @@
 const { app } = require('./firebaseConfig');
-const auth = require('firebase/auth');
 
 const registerUser = async (email, pass) => {
   try {
-    const credentials = await app
-      .auth()
-      .createUserWithEmailAndPassword(email, pass);
-    //const { user } = credentials;
-    console.log('registered. user credentials: ', credentials);
-    return credentials;
+    return await app.auth().createUserWithEmailAndPassword(email, pass);
   } catch (err) {
     console.log(`Unable to register user: ${err.message}`);
     return null;
@@ -17,12 +11,7 @@ const registerUser = async (email, pass) => {
 
 const signInUser = async (email, pass) => {
   try {
-    const credentials = await app
-      .auth()
-      .signInWithEmailAndPassword(email, pass);
-    //const { user } = credentials;
-    console.log('signed in. user credentials: ', credentials);
-    return credentials;
+    return await app.auth().signInWithEmailAndPassword(email, pass);
   } catch (err) {
     console.log(`Unable to log user in: ${err.message}`);
     return null;
