+++ static.json
@@ -1,0 +1,7 @@
+{
+  "root": "public/",
+  "clean_urls": false,
+  "routes": {
+    "/**": "index.html"
+  }
+}
